SA-1849 - Rename components for the sake of consistency
